106 tax lot repository
Move data formatting from tax lot repo to service
